+++ .settings/org.eclipse.core.resources.prefs
@@ -2,6 +2,8 @@ eclipse.preferences.version=1
 encoding//adm_uni/__manifest__.py=utf-8
 encoding//adm_uni/controllers/admission_application_controller.py=utf-8
 encoding//adm_uni/controllers/admission_inquiry_controller.py=utf-8
+encoding//adm_uni/models/__init__.py=utf-8
 encoding//adm_uni/models/admission_application.py=utf-8
 encoding//adm_uni/models/admission_inquiry.py=utf-8
 encoding//adm_uni/models/attachment.py=utf-8
+encoding//adm_uni/models/languages.py=utf-8
