+++ adm_uni/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from . import languages
 from . import attachment
 from . import admission_inquiry
 from . import admission_application
